pyproject.toml
[tool.poetry]
name = "url_shortener"
version = "0.1.0"
description = "A FastAPI-based URL shortener with MongoDB and Beanie ODM"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
beanie = "^1.27.0"
motor = "^3.6.0"
pydantic = "^2.10.1"
python-dotenv = "^1.0.1"
slowapi = "^0.1.8"
redis = "^5.0.4"

[tool.poetry.dev-dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.1"
pre-commit = "^4.3.0"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
